Reference

UnitTestDesign.UnitTestDesign — Module

Generates test cases, which are sets of arguments to use for testing functions.

UnitTestDesign.GND — Type

Greedy Non-deterministic (GND).

This algorithm searches for test cases. It will generate tuples of any order. It generates a different set every time it is invoked.

Arguments

rng::Random.AbstractRNG: This option is a random number generator. Set this if you want to generate the same test cases twice in a row.

M::Int: The number of times it should create a candidate test case each time it creates a candidate. The default is 50. Raising this number could improve test cases and slow generation.

Extended

This algorithm starts with the seeded test cases and then adds test cases, one at a time. It chooses those parameters that are least used, so far, and then chooses values of those parameters that are least covered by previous tuples. At each step, there is some probability of choosing among nearly-equal next values.

It's not complicated, but it can be slow because every next choice checks against all possible tuples. For large numbers of parameters or large numbers of possible values of each parameter, this generator can be slow, so test it for fewer values first and gradually increase the number of parameters or parameter values.

UnitTestDesign.IPOG — Type

In-parameter-order General (IPOG).

This algorithm generates test cases quickly. It will generate tuples of any order. It always generates the same set of test cases for the same set of input values.

Lei, Yu, Raghu Kacker, D. Richard Kuhn, Vadim Okun, and James Lawrence. 2008. “IPOG/IPOG-D: Efficient Test Generation for Multi-Way Combinatorial Testing.” Software Testing, Verification & Reliability 18 (3): 125–48.

UnitTestDesign.MatrixCoverage — Type

Represents covered tuples using a two-dimensional matrix. Each column is another tuple to cover. Each row is a parameter. A zero means this parameter is not part of the tuple. The initial matrix represents all tuples to cover. The remain integer is the number of tuples left to cover. As tuples are covered, they are swapped to the end of the uncovered tuples, and the remain value is decremented.

The arity is an integer array, the same length as the number of parameters. Each member of the array is the number of possible values each parameter can take.

This representation should be slow for lookup of tuples and take lots of memory, but it is a clear representation with which to understand the interface to the data structure.

Base.eltype — Method

The element type of the MatrixCoverage class is determined by the element type of the arity because an element large enough to hold the arity is what defines the minimum possible element type.

UnitTestDesign.add_coverage! — Method

add_coverage!(allc, entry)

The coverage matrix, allc, has row_cnt entries that are considered uncovered, and the rest have been covered already. This adds a new entry, which is a complete set of parameter choices. For every parameter that's covered, this moves those to the end. It returns a new row_cnt which is the number of initial uncovered rows.

UnitTestDesign.all_combinations — Method

all_combinations(arity, n_way)

This represents possible coverage as a matrix, one column per parameter, zero if not used. arity is a list of the number of values for each parameter, and n_way is the order of the combinations, most commonly 2-way.

UnitTestDesign.all_combinations_matrix — Method

allcombinationsmatrix(arity, n_way)

Create matrix coverage that includes all n-way combinations of parameters which can take arity values, where arity is a vector of integers to represent the number of possible values for each parameter, in order.

UnitTestDesign.all_pairs — Method

all_pairs(parameters...; kwargs...)

Ensure that the returned test cases include every pair of parameters at least once.

Examples

all_pairs([1, 2, 3], ["a", "b", "c"], [true, false])

UnitTestDesign.all_triples — Method

all_triples(parameters...; kwargs...)

Ensure that the returned test cases include every combination of three parameters at least once.

UnitTestDesign.all_tuples — Method

all_tuples(parameters...; n_way, engine, disallow, seeds, wayness, Counter)

Given a tuple of parameters, generate all test cases that cover all n_way combinations of those parameters.

Arguments

engine=IPOG(): The engine is IPOG() or GND().

disallow=nothing: The disallow function is a function of the parameters that returns true when that combination should be forbidden.

seeds=[]: is a list of test cases that must be included among those

generated.

wayness is a dictionary that specifies subsets of parameters for which to increase the n_way combinations. For instance, if the combinations are two-way, and you want the third-sixth parameters to be three-way covered, use, wayness = Dict(3 => [[3, 4, 5, 6]]).

Counter=IntThe Counter is an integer type to use for

the computation. It must be large enough to hold the integer number of the parameters.

Examples

all_tuples([1, 2, 3], ["a", "b", "c"], [true, false]; n_way = 2)

UnitTestDesign.all_values — Method

all_values(parameters...; kwargs...)

Ensure that the test cases include every value of every parameter at least once.

See also: all_tuples

UnitTestDesign.argmin_rand — Method

argmin_rand(rng, v)

Given a vector, find the index of the smallest value. If more than one value is the smallest, then randomly choose among the smallest values.

UnitTestDesign.coverage_by_parameter — Method

coverage_by_parameter(allc::MatrixCoverage)

Given a coverage matrix, return how many times each parameter participates in an uncovered tuple. The return value is a vector of integers, where each value is the number of times that parameter appears in any uncovered tuple.

UnitTestDesign.coverage_by_value — Method

coverage_by_value(allc, param_idx)

Given a particular parameter, look at that column of the coverage matrix and return a histogram of how many times each value appears in that column. We want to know which parameter has the most tuples to cover so that we can address it first.

UnitTestDesign.fill_consistent_matches — Method

fill_consistent_matches(mc::MatrixCoverage, entry)

Given an entry that's partially decided, fill in every covering tuple that matches the existing values, in any order.

UnitTestDesign.first_match_for_parameter — Method

first_match_for_parameter(mc::MatrixCoverage, param_idx)

Find the first uncovered tuple for a particular parameter value.

UnitTestDesign.ipog — Method

ipog(arity, n_way)

The arity is an integer array of the number of possible values each parameter can take. The n_way is whether each parameter must appear once in the test suite, or whether each pair of parameters must appear together, or whether each triple must appear together. The wayness can be set as high as the length of the arity.

This uses the in-parameter-order-general algorithm. It will always return the same set of values. It takes less time and memory than most other approaches.

This function represents a test set as a two-dimensional array of the same integer type as the input arity. Each value of the array is either an integer number, from 1 to the arity of that parameter, or it is 0 for what the paper calls a don't-care value.

Lei, Yu, Raghu Kacker, D. Richard Kuhn, Vadim Okun, and James Lawrence.

“IPOG/IPOG-D: Efficient Test Generation for Multi-Way Combinatorial

Testing.” Software Testing, Verification & Reliability 18 (3): 125–48.

UnitTestDesign.match_score — Method

match_score(allc, entry)

If we were to add this entry to the trials, how many uncovered n-tuples would it now cover? allc is the matrix of n-tuples. row_cnt is the first set of rows, representing uncovered tuples. n_way is the length of each tuple. Entry is a set of putative paramter values. Returns an integer number of newly-covered tuples.

This is a place to look into alternative algorithms. We could weight the match score by increasing the score for greater wayness.

UnitTestDesign.matches_from_missing — Method

Given an entry in the test set that has missing values, which are zeros, find any matches that could be created by setting those missing values. Return a new version of the entry.

UnitTestDesign.most_matches_existing — Method

most_matches_existing(mc::MatrixCoverage, existing, param_idx)

The existing vector is a partial test case. Each nonzero entry in this test case is considered as decided. This function then determines, for the next parameter, chosen by paramidx, which value of paramidx, between 1 and its arity, would cover the most tuples. It returns a histogram of how many uncovered tuples could be covered, given the existing choices and a particular value of this parameter.

A value of the param_idx parameter matches with a tuple if its existing choices match at least some part of the tuple and don't disagree with any part of the tuple.

UnitTestDesign.multi_way_coverage — Method

Builds a coverage set where subsets of variables have different wayness.

The all-pairs algorithm is 2-way. All-triples is 3-way. This function takes a base_wayness, which would be 2 or 3 for those examples. Then it takes a dictionary that specifies higher wayness for subsets of parameters.

The wayness is a dictionary from an integer, the wayness, to a set of lists of indices that should have that wayness together. The base_wayness is the n_way for the rest of the variables. It should be less than the other sets of waynesses.

UnitTestDesign.next_multiplicative! — Method

next_multiplicative!(values, arity)

Given a set of values for parameters, this returns the next possible value. When it reaches the end, it cycles back to the start. The first value is all ones. The last value equals the arity.

UnitTestDesign.one_parameter_combinations — Method

one_parameter_combinations(arity, n_way)

Generates all combinations that are nonzero for the last parameter. This is for in-parameter-order generation, where we need only those tuples that end with this column being nonzero.

The construction method is to leave out the given parameter and construct all n_way - 1 tuples. Then copy and paste that once for each possible value of the given parameter.

UnitTestDesign.one_parameter_combinations_matrix — Method

one_parameter_combinations(arity, n_way)

Generates all combinations that are nonzero for the last parameter. This is for in-parameter-order generation, where we need only those tuples that end with this column being nonzero.

The construction method is to leave out the given parameter and construct all n_way - 1 tuples. Then copy and paste that once for each possible value of the given parameter.

UnitTestDesign.parameter_cnt — Method

The number of parameters for this design.

UnitTestDesign.remaining_uncovered — Method

How many tuples have not been covered yet.

UnitTestDesign.remove_combinations! — Method

remove_combinations!(mc::MatrixCoverage, disallow)

Remove all tuples that are disallowed by the disallow function. Input to the function is a vector of parameter indexes. The output to the function is whether they are disallowed, as a boolean. This reduces the size of total tuples in the coverage matrix. It doesn't move them to covered tuples. It deletes them.

UnitTestDesign.total_combinations — Method

total_combinations(arity, n_way)

Given an array of the number of values for each parameter and the level of coverage, return the number of n-tuples required for complete coverage.

UnitTestDesign.tuples_in_trials — Method

tuples_in_trials(trials, n_way)

Find the number of distinct tuples in a set of trials. This uses a different method to count and store those tuples, as a check on other methods. It returns a dictionary where each key is the index of non-zero parameters and the value is the set of all ways those values appear.